+++ src/renderer/Components/WebuiCard/FilterLaunchSettings/ListItemComp.tsx
@@ -14,7 +14,7 @@ type Props = {
   // Item description
   description: string;
   // On check and uncheck value change
-  onValueChange?: (name: {id: string; value: string}, enabled: boolean) => void;
+  onValueChange?: (name: string, enabled: boolean) => void;
 };
 
 function ListItemComp({name, defaultEnabled, description, onValueChange}: Props) {
@@ -38,7 +38,7 @@ function ListItemComp({name, defaultEnabled, description, onValueChange}: Props)
       whileHover={{backgroundColor: isDarkMode ? getBlack(0.3) : getWhiteThird()}}
       ref={scope}
       className="mt-5 flex w-[90%] flex-col items-center justify-center bg-LynxWhiteSecond px-4 py-2 pb-4 dark:bg-black/20">
-      <LCheckBox onValueChange={onValueChange} name={name} defaultEnabled={defaultEnabled} extraClasses="!font-medium" />
+      <LCheckBox onClick={onValueChange} name={name} defaultEnabled={defaultEnabled} extraClasses="!font-medium" />
       <span className="flex text-xl text-black/60 dark:text-white/60">{description}</span>
     </motion.div>
   );